Moves lib to src to make homebrew happy
