+++ src/definition-node.js
