+++ src/iterable-regexp.js
